github-app/azure-function/requirements.txt
azure-functions
PyGithub>=2.1.0
pydantic>=2.0.0
pyyaml>=6.0
cryptography>=42.0.0
